requirements.txt
asyncpg==0.25.0
azure-core==1.23.1
azure-functions==1.10.1
azure-storage-queue==12.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
databases==0.5.5
Flask==2.1.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msrest==0.6.21
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.34
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
